tagger-api using aidr-task-manager

+++ aidr-tagger-api/src/main/java/qa/qcri/aidr/predictui/api/CrisisResource.java
@@ -139,7 +139,7 @@ public class CrisisResource {
         CrisisDTO dto = new CrisisDTO();
         dto.setCode(c.getCode());
         dto.setName(c.getName());
-        dto.setCrisisID(c.getCrisisID());
+        dto.setCrisisID(c.getCrisisID().intValue());
         dto.setCrisisType(typeDTO);
         return dto;
     }
